Fixed a missed step in wiring up the command emitter registration. Also refactored preprocessor registration into the same file - (e.g., register), and updated the code-export-util API accordingly.
